pyproject.toml
[project]
name = "otter"
version = "0.1.0"
description = "An efficient weather forecasting architecture."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cartopy>=0.25.0",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "netcdf4>=1.7.3",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.3.251201",
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "ruff>=0.14.10",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "ty>=0.0.8",
    "types-pyyaml>=6.0.12.20250915",
    "types-tqdm>=4.67.0.20250809",
    "wandb>=0.23.1",
    "weatherbenchx",
    "xarray>=2025.12.0",
    "xformers>=0.0.33.post2",
    "zarr>=3.1.5",
    "apache-beam[gcp,tfrecord]>=2.70.0",
    "dask[array,complete]>=2024.8.0",
    "xarray-beam>=0.11.5",
    "cftime>=1.6.5",
    "pyarrow>=18.1.0",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
weatherbenchx = { git = "https://github.com/google-research/weatherbenchX.git" }

[tool.ruff]
line-length = 79
lint.extend-select = ["RUF012", "RUF100", "RUF200"]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.ty.rules]
unresolved-attribute = "ignore"
